requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2023.12.12=hecd8cb5_0
docopt=0.6.2=pypi_0
imageio=2.34.0=pypi_0
joblib=1.3.2=pypi_0
lazy-loader=0.3=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
ncurses=6.4=hcec6c5f_0
networkx=3.2.1=pypi_0
nptyping=2.5.0=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.13=hca72f7f_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39hecd8cb5_0
pykwalify=1.8.0=pypi_0
pynrrd=1.0.0=pypi_0
pyradiomics=3.0.1=pypi_0
python=3.9.18=h5ee71fb_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.5.0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=hca72f7f_0
ruamel-yaml=0.18.6=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
setuptools=68.2.2=py39hecd8cb5_0
shapely=2.0.3=pypi_0
simpleitk=2.3.1=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h6c40b1e_0
threadpoolctl=3.3.0=pypi_0
tifffile=2024.2.12=pypi_0
tk=8.6.12=h5d9f67b_0
typing-extensions=4.10.0=pypi_0
tzdata=2024.1=pypi_0
wheel=0.41.2=py39hecd8cb5_0
xz=5.4.6=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
